fix: cleanup and rename modelsingleselectlegacy
