+++ src/components/metadataFormControls/ModelSingleSelect/BaseModelSingleSelect.tsx
@@ -1,3 +1,4 @@
+import i18n from '@dhis2/d2-i18n'
 import React, { useCallback, useMemo } from 'react'
 import { DisplayableModel } from '../../../types/models'
 import {
@@ -5,16 +6,17 @@ import {
     SearchableSingleSelectPropTypes,
 } from '../../SearchableSingleSelect'
 
+
 const toDisplayOption = (model: DisplayableModel) => ({
     value: model.id,
-    label: model.displayName,
+    label: model.displayName || i18n.t('Loading...'),
 })
 
 type OwnProps<TModel> = {
     selected?: TModel
     available: TModel[]
     onChange: (selected: TModel | undefined) => void
-    showNoValueOption?: boolean
+    noValueOption?: { value: string; label: string } | boolean
 }
 
 export type BaseModelSingleSelectProps<TModel extends DisplayableModel = DisplayableModel> = Omit<
@@ -28,7 +30,7 @@ export const BaseModelSingleSelect = <TModel extends DisplayableModel>({
     available,
     selected,
     onChange,
-    showNoValueOption,
+    noValueOption,
     ...searchableSingleSelectProps
 }: BaseModelSingleSelectProps<TModel>) => {
     const { allModelsMap, allSingleSelectOptions } = useMemo(() => {
@@ -40,15 +42,16 @@ export const BaseModelSingleSelect = <TModel extends DisplayableModel>({
         const allSingleSelectOptions = Array.from(allModelsMap).map(
             ([, value]) => toDisplayOption(value)
         )
-        if (showNoValueOption) {
-            allSingleSelectOptions.unshift({ value: '', label: '' })
+        if (noValueOption) {
+            const option = noValueOption === true ? { value: '', label: i18n.t('<No value>') } : noValueOption
+            allSingleSelectOptions.unshift(option)
         }
 
         return {
             allModelsMap,
             allSingleSelectOptions,
         }
-    }, [available, selected, showNoValueOption])
+    }, [available, selected, noValueOption])
 
     const handleOnChange: SearchableSingleSelectPropTypes['onChange'] =
         useCallback(
