+++ src/components/metadataFormControls/ModelSingleSelect/index.ts
@@ -1,1 +1,4 @@
+export * from './ModelSingleSelectLegacy'
+export * from './useRefreshSingleSelect'
+export * from './ModelSingleSelectField'
 export * from './ModelSingleSelect'
