+++ src/pages/dataElements/fields/CategoryComboField.tsx
@@ -1,13 +1,17 @@
 import i18n from '@dhis2/d2-i18n'
+import { Field } from '@dhis2/ui'
 import React, { useEffect } from 'react'
 import { useField, useForm } from 'react-final-form'
 import { useHref } from 'react-router'
-import { EditableFieldWrapper } from '../../../components'
-import { ModelSingleSelectField, ModelSingleSelectFieldProps } from '../../../components/metadataFormControls/ModelSingleSelect/ModelSingleSelectField'
+import { CategoryComboSelect, EditableFieldWrapper } from '../../../components'
+import {
+    ModelSingleSelectFieldProps
+} from '../../../components/metadataFormControls/ModelSingleSelect/ModelSingleSelectField'
 import { useRefreshModelSingleSelect } from '../../../components/metadataFormControls/ModelSingleSelect/useRefreshSingleSelect'
 import { DEFAULT_CATEGORY_COMBO } from '../../../lib'
 import { PlainResourceQuery } from '../../../types'
 import classes from './CategoryComboField.module.css'
+import { DisplayableModel } from '../../../types/models'
 
 const query = {
     resource: 'categoryCombos',
@@ -18,7 +22,9 @@ const query = {
 } as const satisfies PlainResourceQuery
 
 // stable reference for transform function
-const withDefaultCategoryCombo: ModelSingleSelectFieldProps['transform'] = (value) => [DEFAULT_CATEGORY_COMBO, ...value]
+const withDefaultCategoryCombo: ModelSingleSelectFieldProps['transform'] = (
+    value
+) => [DEFAULT_CATEGORY_COMBO, ...value]
 /*
  * @TODO: Verify that the api ignores the category combo when it's disabled.
  *        If it does not, file a jira issue and "escalate" this so it will be
@@ -34,7 +40,7 @@ export function CategoryComboField() {
     const {
         input: { value: domainTypeValue },
     } = useField('domainType')
-
+    const { input, meta } = useField<DisplayableModel>('categoryCombo', { validateFields: [] })
     const domainTypeIsTracker = domainTypeValue === 'TRACKER'
     const disabled = domainTypeIsTracker
     const newCategoryComboLink = useHref('/categoryCombos/new')
@@ -52,14 +58,39 @@ export function CategoryComboField() {
             onAddNew={() => window.open(newCategoryComboLink, '_blank')}
         >
             <div className={classes.categoryComboSelect}>
-                <ModelSingleSelectField
-                    query={query}
-                    name="categoryCombo"
-                    label={i18n.t('Category combo')}
+                <Field
                     required
-                    transform={withDefaultCategoryCombo}
-                    disabled={disabled}
-                />
+                    name="categoryCombo"
+                    label={i18n.t('{{fieldLabel}} (required)', {
+                        fieldLabel: i18n.t('Category combination'),
+                    })}
+                    helpText={i18n.t(
+                        'Choose how this data element is disaggregated'
+                    )}
+                    error={meta.touched && !!meta.error}
+                    validationText={meta.touched ? meta.error : undefined}
+                    dataTest="formfields-categorycombo"
+                >
+                    <CategoryComboSelect
+                        selected={input.value}
+                        disabled={disabled}
+                        onChange={(selected) => {
+                            input.onChange(selected)
+                            input.onBlur()
+                        }}
+                        onBlur={input.onBlur}
+                        onFocus={input.onFocus}
+                        query={{
+                            resource: 'categoryCombos',
+                            params: {
+                                filter: [
+                                    'isDefault:eq:false',
+                                    'dataDimensionType:eq:DISAGGREGATION',
+                                ],
+                            },
+                        }}
+                    />
+                </Field>
             </div>
         </EditableFieldWrapper>
     )
